Supports dynamic date range for fetching closed pull requests.
Currently, while dynamic, the date range will always be the last week.
Eventually, we could add some customization around this (e.g., passing
in start or end date, etc.).

A big part of this was turning the GitHub module into a class, to make
mocking and dependency injection easier. By changing it to a class, and
having `index.ts` take it as an argument with a default, we were able to
write tests that pass in a fake version (see `FakeService` in
`idnex.test.ts`). The fake version could have the API call mocked out,
making testing the date stuff easier.

Moving away from a hard coded date to a dyanmic date range, I learned
the syntax for a date range in the GitHub search is `date1..date2` (as
opposed to something like `closed:>date1 closed:<date2`).

This commit also contains a small tweak to `tsconfig`.

-    // "lib": [],                             /* Specify library files to be included in the compilation. */
+    "lib": [
+      "es2015"

To be honest, not 100% why this was necessary, but it was the
recommendation for compiler error.