Changes how the graphql body is tested.
Leverages mock fetch to test the posted body to see the graphql body being sent in `Github.recentlyClosedPullRequets`.

Previously, this was tested by using the extracted function, `GutHub.searchQueryString`. Testing the posted body feels like it probably tests more of what we ultimately want to test. Really, we wanted `searchQueryString` to be a private helper function. This commit prefixes the function with `__` to indicate it's private to others.