Updates deploy filtering on CircleCI
Based on [this
thread](https://discuss.circleci.com/t/workflow-job-with-tag-filter-being-run-for-every-commit/20762/3)
it seems filters are ORed. This commit reverts the filter back to how it
was. Basically, this means that it will ignore all branches unless there
is a tag.

I was hoping to go for a workflow of `master` AND `tag`, but it seems
tis may be non-standard.

Instead, version bumping will probably need to be somewhat manually
done, but once the version is bumped and push, things should get tested
and published.