refactor(github): changes setup for GitHub auth token
Previously, we expected a `GITHUB_AUTH_TOKEN` environment variable to be
set. To make things more flexible, this commit changes to require you to
make an instance of a `GitHub` class, and pass in the auth token. This
allows the end user to get the token however they want and no longer
ties them to specific environment variables.

BREAKING CHANGE: Previously, you had to have a `GITHUB_AUTH_TOKEN`
environment variable that would be used in `GitHub` class methods. The
API has changed to instead create in instance of `GitHub` by passing in
your auth token however works for you in your code. Previously used
class method `recentlyClosedPullRequests` is now an instance method on
`GitHub`.