refactor(index.ts): updates `index` to use instance an of a GitHub class
- Updates `fetchRecentlyClosedPullRequests` in `index` to use the new
version of `GitHub` that requires creating an instance that is passed in
the auth token.
- Updates `fetchRecentlyClosedPullRequests` to require an `auth_token`
is passed in
- Updates tests for `fetchRecentlyClosedPullRequests` to use different
jest mocking functionality and mock out the implementation of `GitHub`.
Now, instead of passing in a fake module that matches `GitHub` we use
magic jest mocking to create a mocked GitHub import which lets us mock
the constructor.

BREAKING CHANGE: Updates the function `fetchRecentlyClosedPullRequests`
exported from `index.ts` to now take in the auth token. This allows for
more flexibilty in that the auth token no longer **has** to be fetched
from an environment variable named `GIT